Enable bf16 SSM state in fused GDN kernels by default on TPU (#2482)

+++ tpu_inference/kernels/gdn/fused_gdn_kernel_wrapper.py
@@ -57,12 +57,16 @@ def _dispatch_with_distribution(
     before it can decode), so masking is unnecessary on the decode path.
     """
     # ── Decode kernel → updates state in-place ──
+    # NOTE: pass `initial_state` through as-is. The kernels upcast to fp32
+    # on VMEM load for compute precision; HBM storage stays at the array's
+    # dtype. An `astype(jnp.float32)` here would materialize an fp32 copy
+    # of a bf16 state and undo the storage win before the kernel runs.
     o_d, state_1 = fused_decoding_gdn(
         q,
         k,
         v,
         g.astype(jnp.float32),
-        initial_state.astype(jnp.float32),
+        initial_state,
         state_indices,
         distribution,
         b,
